@@ -17,23 +17,34 @@
     "test": "jest --config ./config/jest/config.json --no-cache",
     "test:coverage": "npm run test -- --coverage",
     "test:watch": "npm run test -- --watch",
-    "test:ci": "npm run lint && npm run test && codecov"
+    "test:ci": "npm run lint && npm run test && codecov",
+    "check:format": "prettier --check '**/*.{ts,tsx,js,jsx}' '!**/dist/**' '!**/coverage/**'"
   },
   "dependencies": {
-    "prop-types": "^15.5.7"
+    "@babel/preset-env": "^7.18.10",
+    "core-js": "^3.24.1",
+    "prettier": "^2.7.1",
+    "prop-types": "^15.5.7",
+    "react-scripts": "3.3.0",
+    "ts-jest": "^28.0.7"
   },
   "peerDependencies": {
-    "react": "15.x || 16.x"
+    "react": "15.x || 16.x || 17.x"
   },
   "devDependencies": {
-    "@types/jest": "^20.0.8",
+    "@testing-library/dom": "^8.16.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "12.0.0",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/user-event": "^14.4.2",
+    "@types/jest": "^28.1.6",
     "@types/prop-types": "^15.5.1",
-    "@types/react": "^15.0.38",
-    "@types/react-dom": "^15.5.4",
+    "@types/react": "17.0.13",
+    "@types/react-dom": "17.0.8",
     "awesome-typescript-loader": "^3.2.1",
     "babel-cli": "^6.26.0",
     "babel-eslint": "7.0.0",
-    "babel-jest": "^19.0.0",
+    "babel-jest": "^28.1.3",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-typescript": "^7.0.0-alpha.19",
@@ -41,11 +52,12 @@
     "eslint": "^4.10.0",
     "eslint-plugin-shopify": "^22.1.0",
     "fs-extra": "^4.0.1",
-    "jest": "^19.0.2",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
     "nwb": "0.15.x",
-    "react": "^15.4.2",
+    "react": "17.0.2",
     "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react-dom": "17.0.2",
     "react-test-renderer": "^15.4.2",
     "rimraf": "^2.6.1",
     "rollup": "^0.49.2",
@@ -56,7 +68,7 @@
     "rollup-plugin-uglify": "^2.0.1",
     "tslint": "^5.10.0",
     "tslint-config-shopify": "^3.0.2",
-    "typescript": "2.8.3"
+    "typescript": "^4.7.4"
   },
   "author": {
     "name": "Clauderic Demers",
