@@ -1,3 +1,3 @@
 module.exports = require('babel-jest').createTransformer({
-  presets: ['es2015', 'react', 'stage-1'],
+  presets: ['@babel/preset-env'],
 });
