@@ -1,11 +1,11 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import * as React from "react";
+import * as ReactDOM from "react-dom";
 
-import VirtualList, {ItemStyle} from '../../src';
-import './demo.css';
+import VirtualList, { ItemStyle } from "../../src";
+import "./demo.css";
 
 class Demo extends React.Component {
-  renderItem = ({style, index}: {style: ItemStyle; index: number}) => {
+  renderItem = ({ style, index }: { style: ItemStyle; index: number }) => {
     return (
       <div className="Row" style={style} key={index}>
         Row #{index}
@@ -29,4 +29,4 @@ class Demo extends React.Component {
   }
 }
 
-ReactDOM.render(<Demo />, document.querySelector('#app'));
+ReactDOM.render(<Demo />, document.querySelector("#app"));
